requirements.txt
asgiref==3.2.7
astroid==2.4.0
autopep8==1.5.2
certifi==2020.4.5.1
colorama==0.4.3
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.0.3
django-cors-headers==3.2.1
django-heroku==0.3.1
django-import-export==2.1.0
django-monthfield==0.1.3
djangorestframework==3.11.0
et-xmlfile==1.0.1
gunicorn==20.0.4
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.2
psycopg2==2.8.5
pycodestyle==2.5.0
pylint==2.5.0
pytz==2020.1
PyYAML==5.3.1
sqlparse==0.3.1
tablib==2.0.0
toml==0.10.0
whitenoise==4.1.4
wincertstore==0.2
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
